+++ NEWS.md
@@ -1,6 +1,8 @@
 
 # renv 0.16.0  (UNRELEASED)
 
+* `renv` now handles local sources within the current working directory. (#1079)
+
 * The `renv` sandbox is once again enabled by default. On Unix, the sandbox
   is now created by default within the project's `renv/sandbox` library.
   On Windows, the sandbox is created within the R session's `tempdir()`.